+++ docs/CHANGELOG.md
@@ -17,6 +17,8 @@ and this programmer does her best to follow [Conventional Commits](https://www.c
 
 ### Changed
 
+- Now uses the [perflogger package](https://github.com/ewigael/perflogger) for performance tracking
+
 ### Removed
 
 ### Fixed